requirements.txt
fastapi
openai
python-docx
uvicorn
python-multipart
langchain
langchain-core
langchain-openai
langchain-community
langchain-chroma
